# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Body:
    properties:
      data: {}
      error:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  setting.SetConfigRequest:
    properties:
      path:
        type: string
      scope:
        type: string
      scope_id:
        type: string
      value:
        type: string
    required:
    - path
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Modular Go + Gin Backend API (REST & GraphQL)
  title: CodebaseGo API
  version: "1.0"
paths:
...
  /settings:
    get:
      parameters:
      - description: Filter by path prefix (e.g. web/)
        in: query
        name: path_prefix
        type: string
      - description: Filter by scope
        in: query
        name: scope
        type: string
      - description: Opaque cursor for pagination
        in: query
        name: cursor
        type: string
      - default: 10
        description: Items per page for cursor pagination
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page number for offset pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page for offset pagination
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: List system configurations
      tags:
      - settings
    post:
      consumes:
      - application/json
      parameters:
      - description: Config payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/setting.SetConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Set configuration value
      tags:
      - settings
  /settings/{id}:
    delete:
      parameters:
      - description: Config ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Delete config by ID
      tags:
      - settings
    get:
      parameters:
      - description: Config ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Get config by ID
      tags:
      - settings
  /settings/by-path:
    get:
      parameters:
      - description: Config Path (e.g. web/unsecure/base_url)
        in: query
        name: path
        required: true
        type: string
      - default: default
        description: Scope
        in: query
        name: scope
        type: string
      - default: "0"
        description: Scope ID
        in: query
        name: scope_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      summary: Get config by path
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"